Refactor: code review fixes — security, DRY, async, and modular structure
- Move SMTP credentials (user + password) to macOS Keychain
- Sanitize error messages to strip filesystem paths
- Add TTL caching for contacts/reminders DB paths
- Validate dates in toCoreDataTimestamp (throw on invalid input)
- Set temp file permissions to 0o600
- Add logging to silent catch blocks
- Improve UUID regex in FTS index
- Fix "Query too long" errors on non-query fields
- Extract shared readEmlxRaw/readEmlxRawAsync helpers (DRY)
- Consolidate mailbox URL filter into shared config
- Make getBodyPreview async to avoid blocking event loop
- Split index.ts (1166 lines) into per-domain register.ts files (256 lines)
- Add shared mcp-helpers.ts for ok/err/toMarkdown utilities

All 143 tests pass.
